requirements.txt
attrs==19.3.0             # via pytest
certifi==2020.4.5.1       # via requests
chardet==3.0.4            # via requests
django==3.0.5             # via -r requirements.in, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
idna==2.9                 # via requests
importlib-metadata==1.6.0  # via pluggy, pytest
more-itertools==8.2.0     # via pytest
packaging==20.3           # via pytest
pluggy==0.13.1            # via pytest
py==1.8.1                 # via pytest
pyparsing==2.4.7          # via packaging
pytest-django==3.9.0      # via -r requirements.in
pytest==5.4.1             # via pytest-django
requests==2.23.0          # via -r requirements.in
six==1.14.0               # via packaging
sqlparse==0.3.1           # via django
urllib3==1.25.9           # via requests
wcwidth==0.1.9            # via pytest
zipp==3.1.0               # via importlib-metadata
isort
python-slugify
djangorestframework-simplejwt
django-filter~=2.4.0
PyJWT~=2.0.0
colorama~=0.4.4
pip~=20.2.1
setuptools~=49.2.1
numpy~=1.19.5
atomicwrites~=1.4.0
python-dotenv~=0.15.0
django-cors-headers

asgiref==3.2.10
gunicorn==20.0.4
psycopg2-binary
pytz==2020.1
sqlparse==0.3.1
